% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_means.R
\name{od_means}
\alias{od_means}
\title{ODMeans}
\usage{
od_means(data, numK, limitsSeparation, maxDist, distHierarchical)
}
\arguments{
\item{data}{A data frame with four columns:
Initial Latitude | Initial Longitude | Final Latitude | Final Longitude}

\item{numK}{Initial number of clusters in the first call of K-Means.}

\item{limitsSeparation}{Range to determine if a drastic change has happened between a cluster and its separation.
A bigger value makes more difficult to separate a cluster.}

\item{maxDist}{Maximum distance to join two points. This is based on the euclidean distance.}

\item{distHierarchical}{Maximum distance to create a new hierarchy per cluster}
}
\value{
Returns a structure that contains the final centers, clusters, sizes and hierarchy
}
\description{
ODMeans
}
\examples{
data(ODMeansSampleData)
od_means(ODMeansSampleData, 5, 200, 2500, 500)
}
