% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMeansSampleData.R
\docType{data}
\name{ODMeansSampleData}
\alias{ODMeansSampleData}
\title{Origin-Destination points}
\format{
A data frame with 10000 rows and 4 variables:
\describe{
  \item{OriginLatitude}{Consists of the origin latitude dimension}
  \item{OriginLongitude}{Consists of the origin longitude dimension}
  \item{DestinationLatitude}{Consists of the destination latitude dimension}
  \item{DestinationLongitude}{Consists of the destination longitude dimension}
  \item{original_cluster}{Original cluster of the points when it was created}
  ...
}
}
\source{
Synthetic data
}
\usage{
ODMeansSampleData
}
\description{
A dataset containing 1700 Origin-Destination points
}
\keyword{datasets}
