% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMeansTaxiData.R
\docType{data}
\name{ODMeansTaxiData}
\alias{ODMeansTaxiData}
\title{Origin-Destination Taxi data}
\format{
A data frame with 452,166 rows and 4 variables:
\describe{
  \item{OriginLatitude}{Origin latitude dimension}
  \item{OriginLongitude}{Origin longitude dimension}
  \item{DestinationLatitude}{Destination latitude dimension}
  \item{DestinationLongitude}{Destination longitude dimension}
  ...
}
}
\source{
Fantaxico, Fermanti Servicios de Ingeniería S.A, Santiago, Chile.
}
\usage{
data(ODMeansTaxiData)
}
\description{
The data frame contains the 452,166 trips collected for the months of March (2014 to 2016), July (2014 to 2016), and November (2014 and 2015). The data points are the taxi's initial and ending location based on latitude and longitude.
}
\keyword{datasets}
